+++ app/src/main/java/com/example/wbproject/ui/theme/screens/more/my_meetings/MyMeetingScreen.kt
@@ -32,7 +32,7 @@ import org.koin.androidx.compose.koinViewModel
     ExperimentalFoundationApi::class
 )
 @Composable
-fun MyMeetingScreen(modifier: Modifier = Modifier, onMeetingCardClickListener: () -> Unit) {
+fun MyMeetingScreen(modifier: Modifier = Modifier, onMeetingCardClickListener: (Int) -> Unit) {
     val viewModel: MyMeetingViewModel = koinViewModel()
     val myMeetingState by viewModel.getMyMeetingFlow().collectAsState()
     val pagerState = rememberPagerState(pageCount = { TabsForMyMeetingList.entries.size })
